+++ models/comparison_methods/lstm.py
@@ -14,7 +14,7 @@ from torch import nn
 from typing import Dict
 
 
-class LSTMNet(nn.Module):
+class LSTM_Net(nn.Module):
     """ The 2 Layer LSTM. hidden_size=64. """
 
     def __init__(
@@ -31,23 +31,21 @@ class LSTMNet(nn.Module):
 
         """
 
-        super(LSTMNet, self).__init__()
+        super(LSTM_Net, self).__init__()
         self.device = device
-        self.input_size = input_size
 
         # ---- Log the info of LSTM ---- #
-        logging.info(f"|||| Using LSTM Now ! input_size={input_size}, hidden_size={hidden_size}, "
-                     f"num_layers={num_layers}, dropout_ratio={dropout}||||")
+        logging.info(f"|||| Using LSTM Now ! input_size={input_size}, hidden_size={hidden_size}, num_layers={num_layers}, dropout_ratio={dropout}||||")
 
-        # ---- Part 1. The rnn of the LSTM module ---- #
-        self.rnn = nn.LSTM(
+        # ---- Part 1. The LSTM module ---- #
+        self.lstm = nn.LSTM(
             input_size=input_size, hidden_size=hidden_size, num_layers=num_layers, batch_first=True, dropout=dropout
         ).to(device=device)
 
         # ---- Part 2. The output fully connect layer ---- #
         self.fc_out = nn.Linear(hidden_size, 1).to(device=device)
 
-    def forward(self, mul_granularity_input: Dict[str, torch.Tensor]):
+    def forward(self, mul_granularity_input: Dict[str, torch.Tensor]) -> Dict[str, torch.Tensor]:
         """ The forward function of LSTM Net.
 
         :param mul_granularity_input: the input multi granularity, a dict with the format:
@@ -63,17 +61,17 @@ class LSTMNet(nn.Module):
 
         """
 
-        # ---- Step 1. Get the coarsest feature ---- #
+        # ---- Step 1. Get the feature ---- #
         # g1 feature (coarsest), shape=(bs, T, K^g1, D)
         feature_g1 = mul_granularity_input["g1"].to(dtype=torch.float32, device=self.device)
         # get the feature shape
-        bs, T = feature_g1.shape[0], feature_g1.shape[1]
+        bs, T, K_g1, d = feature_g1.shape[0], feature_g1.shape[1], feature_g1.shape[2], feature_g1.shape[3]
 
-        # ---- Step 2. Reshape the input for encoding ---- #
-        feature_g1 = feature_g1.reshape(bs, T, self.input_size)  # shape=(bs, T, K^g1*D)
+        # ---- Step 2. Preprocess the input for encoding ---- #
+        feature_g1 = feature_g1.reshape(bs, T, K_g1 * d)  # reshape, shape=(bs, T, K^g1*D)
 
-        # ---- Step 3. RNN Encoding and get the hidden_feature ---- #
-        hidden_g1, _ = self.rnn(feature_g1)  # shape=(bs, T, hidden_size)
+        # ---- Step 3. LSTM Encoding and get the hidden_feature ---- #
+        hidden_g1, _ = self.lstm(feature_g1)  # shape=(bs, T, hidden_size)
 
         # ---- Step 4. FC to get the prediction ---- #
         # get the last step hidden g1
@@ -98,6 +96,6 @@ if __name__ == "__main__":  # A demo of LSTM
     g_dict = {"g1": 1, "g2": 2, "g3": 6, "g4": 24, "g5": 96}
     dev = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
-    model = LSTMNet(input_size=1, device=dev)
+    model = LSTM_Net(input_size=1, device=dev)
     out = model(mg_input)
     print(out["pred"].shape)